avengers: Replace protection simulation with hookup of emulated MCU; add wait states to main CPU to keep it in sync while downloading the palette from the MCU [AJR, David Haywood]
* lwings.cpp: Merge driver files; avoid runtime memory bank tag lookups
